+++ crates/core/src/editor_selection.rs
@@ -13,7 +13,7 @@ mod editor_selection_transform;
 use self::editor_selection_state::VisualSelection;
 use self::editor_selection_transform::{
     CamelCaseTransform, KebabCaseTransform, PascalCaseTransform, ScreamingSnakeCaseTransform,
-    SnakeCaseTransform,
+    SnakeCaseTransform, TrainCaseTransform,
 };
 
 pub(super) struct EditorSelection {
@@ -116,6 +116,19 @@ impl EditorSelection {
     ) {
         let range = self.visual.range_for(cursor, file, line_view);
         let transform = PascalCaseTransform;
+        range.apply_to(file, line_view, cursor_update, &transform);
+        self.visual.clear();
+    }
+
+    pub(super) fn train(
+        &mut self,
+        cursor: CursorPosition,
+        file: &mut File,
+        line_view: &LineView,
+        cursor_update: &mut CursorUpdate<'_>,
+    ) {
+        let range = self.visual.range_for(cursor, file, line_view);
+        let transform = TrainCaseTransform;
         range.apply_to(file, line_view, cursor_update, &transform);
         self.visual.clear();
     }