+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Conventional Commits.
 - snake_case transformation command for visual selections via `:case snake`.
 - SCREAMING_SNAKE_CASE transformation command for visual selections via `:case screaming`.
 - PascalCase transformation command for visual selections via `:case pascal`.
+- Train-Case transformation command for visual selections via `:case train`.
 
 ## [0.1.0] - 2025-12-30